Move notification API calls into dedicated module
